edited data fields

+++ db/migrate/20220913211714_add_price_to_books.rb
@@ -1,5 +1,5 @@
 class AddPriceToBooks < ActiveRecord::Migration[6.1]
   def change
-    add_column :books, :price, :float
+    add_column :books, :price, :decimal
   end
 end
